backend/app/requirements.txt
alembic==1.13.3
fastapi==0.115.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-multipart==0.0.9
SQLAlchemy==2.0.43
uvicorn==0.30.6
